fix: check for libp2p key in name pubsub subs
The list of ipns-over-pubsub subscription has topic names that contain
the peer id as a libp2p key, not a base58btc encoded bare multihash,
so convert the enocded multihash to a libp2p key first, otherwise
it says it can't find the key in the subs list when it is there.

Also uses `uint8arrays/to-string` to format pubsub messages in hex
form, just calling `.toString()` on it won't work as the API is
different to node Buffers.